refactor(erational): split erational into core / manipulators / iostream (#1334)
    core.hpp          arithmetic, no streams          74,298 lines, 0 I/O-family
    manipulators.hpp  type_tag / to_string
    iostream.hpp      operator<< / operator>>
    erational.hpp     everything, as before           82,501 lines

The umbrella was 101,296 lines pulling all five stream headers, so the core is 27%
smaller and pulls none -- the largest reduction of the rollout.

Two of those savings were not layering at all:

<regex> was included and never used. parse() splits on '/' with find/substr and defers to
scan_decimal_float; nothing in the file constructs a regex. libstdc++'s <regex> pulls
<sstream>, so the dead include alone cost 19,000 lines and two of the five I/O headers.

erational_impl.hpp included edecimal's UMBRELLA rather than its core. erational stores its
numerator and denominator as edecimal, which is a structural dependency on the arithmetic,
not on edecimal's text layer. Re-pointed at edecimal/core.hpp (#1423). erational/iostream.hpp
correspondingly includes edecimal/iostream.hpp, because erational's operator<< streams
those two members through edecimal's.

The impl now takes ieee754_core.hpp and manipulators_core.hpp for extractFields,
ieee754_parameter and scale. Three erational_divide_by_zero diagnostics became
fprintf(stderr, ...); the friend declarations need only <iosfwd>.

type_tag(const erational&) was a non-template free function defined in a header WITHOUT
inline -- an ODR violation, not a style point: any two translation units that both
included erational.hpp failed to link on it. It is now inline.

That fix is partial and the rest is NOT addressed here. erational still has 68 duplicate
symbols across two translation units, and edecimal 12, all pre-existing on main with
identical counts (measured before and after). They are the whole erational mathlib -- sin,
cos, exp, log, floor, ... -- plus ltrim/rtrim in string/strmanip.hpp and the edecimal
comparison and division helpers. Filed separately rather than folded into a refactor.

Also closed the IWYU gaps the hardened #1389 sweep found across the subtree: <string> in
exceptions.hpp, <cstdint>/<limits> in numeric_limits.hpp, <type_traits>/<utility> in the
impl, and the <cmath>/<algorithm>/<cstddef> that nine math headers had been getting
transitively. Sweep is clean at 0 gaps across all 24 headers.

Verified: the core gate passes on gcc and clang; all 13 translation units in the tree that
mention erational build and run clean on both; every api.cpp in the tree, 55 pass 0 fail,
on both; ASCII clean.
